+++ src/oceanpack/app/controllers/data_controller.py
@@ -60,7 +60,7 @@ class DataProcessingController:
         self.model.compute_equilibrator_pressure()
         self.model.compute_pCO2_wet_equ()
         self.model.compute_temperature_correction()
-        self.model.compute_fCO2_wet_sst()
+        self.model.compute_fCO2()
 
     def generate_output(self, path):
         self.model.to_netcdf(path)